Plot visualization plot from a ColocBoost output.

colocboost_plot generates visualization plots for colocalization events from a ColocBoost analysis.

Usage

colocboost_plot(
  cb_output,
  y = "log10p",
  pos = NULL,
  plot_target_only = FALSE,
  plot_cos_idx = NULL,
  outcome_idx = NULL,
  points_color = "grey80",
  cos_color = NULL,
  add_vertical = FALSE,
  add_vertical_idx = NULL,
  outcome_names = NULL,
  plot_cols = 2,
  variant_coord = FALSE,
  show_hits = FALSE,
  show_cos_to_uncoloc = FALSE,
  show_cos_to_uncoloc_idx = NULL,
  show_cos_to_uncoloc_outcome = NULL,
  gene_name = NULL,
  ylim_each = TRUE,
  outcome_legend_pos = "top",
  outcome_legend_size = 1.2,
  cos_legend_pos = "bottomleft",
  show_variable = FALSE,
  lab_style = c(2, 1),
  axis_style = c(2, 1),
  title_style = c(2.5, 2),
  ...
)

Arguments

cb_output

Output object from colocboost analysis

y

Specifies the y-axis values, default is "log10p" for -log10 transformed marginal association p-values.

pos

Optional plotting range of x-axis to zoom in to a specific region.

plot_target_only

Logical, if TRUE only plots colocalization with target outcome, default is FALSE.

plot_cos_idx

Optional indices of CoS to plot

outcome_idx

Optional indices of outcomes to include in the plot. outcome_idx=NULL to plot only the outcomes having colocalization.

points_color

Background color for non-colocalized variables, default is "grey80".

cos_color

Optional custom colors for CoS.

add_vertical

Logical, if TRUE adds vertical lines at specified positions, default is FALSE

add_vertical_idx

Optional indices for vertical lines.

outcome_names

Optional vector of outcomes names for the subtitle of each figure. outcome_names=NULL for the outcome name shown in data_info.

plot_cols

Number of columns in the plot grid, default is 2. If you have many colocalization. please consider increasing this.

variant_coord

Logical, if TRUE uses variant coordinates on x-axis, default is FALSE. This is required the variable names including position information.

show_hits

Logical, if TRUE shows top variables for each CoS, default is FALSE

show_cos_to_uncoloc

Logical, if TRUE shows colocalization to uncolocalized outcomes to diagnose, default is FALSE

show_cos_to_uncoloc_idx

Optional indices for showing CoS to all uncolocalized outcomes

show_cos_to_uncoloc_outcome

Optional outcomes for showing CoS to uncolocalized outcomes

gene_name

Optional gene name to display in plot title

ylim_each

Logical, if TRUE uses separate y-axis limits for each plot, default is TRUE

outcome_legend_pos

Position for outcome legend, default is "top"

outcome_legend_size

Size for outcome legend text, default is 1.2

cos_legend_pos

Position for colocalization set legend, default is "bottomleft"

show_variable

Logical, if TRUE displays variant IDs, default is FALSE

lab_style

Vector of two numbers for label style (size, boldness), default is c(2, 1)

axis_style

Vector of two numbers for axis style (size, boldness), default is c(2, 1)

title_style

Vector of two numbers for title style (size, boldness), default is c(2.5, 2)

...

Additional parameters passed to plot functions

Value

Visualization plot for each colcoalization event.

Examples

colocboost_plot(cb_output)
#> Error: object 'cb_output' not found