Get colocalization summary table from a ColocBoost output.

get_cos_summary get the colocalization summary table with or without the outcomes of interest.

Usage

get_cos_summary(
  cb_output,
  outcome_names = NULL,
  interest_outcome = NULL,
  gene_name = NULL
)

Arguments

cb_output

Output object from colocboost analysis

outcome_names

Optional vector of names of outcomes, which has the same order as Y in the original analysis.

interest_outcome

Optional vector specifying a subset of outcomes from outcome_names to focus on. When provided, only colocalization events that include at least one of these outcomes will be returned.

gene_name

Optional character string. When provided, adds a column with this gene name to the output table for easier filtering in downstream analyses.

Value

A summary table for colocalization events with the following columns:

target_outcome

The target outcome being analyzed if exists. Otherwise, it is FALSE.

colocalized_outcomes

Colocalized outcomes for colocalization confidence set (CoS)

cos_id

Unique identifier for colocalization confidence set (CoS)

purity

Minimum absolute correlation of variables with in colocalization confidence set (CoS)

top_variable

The variable with highest variant colocalization probability (VCP)

top_variable_vcp

Variant colocalization probability for the top variable

cos_npc

Normalized probability of colocalization

min_npc_outcome

Minimum normalized probability of colocalized traits

n_variables

Number of variables in colocalization confidence set (CoS)

colocalized_index

Indices of colocalized variables

colocalized_variables

List of colocalized variables

colocalized_variables_vcp

Variant colocalization probabilities for all colocalized variables

Examples

get_cos_summary(cb_output)
#> Error: object 'cb_output' not found
get_cos_summary(cb_output, interest_outcome = c("Y1", "Y2"))
#> Error: object 'cb_output' not found